requirements-dev.txt
python-dotenv
langchain==0.2.6
openai==1.35.7
langchain-openai==0.1.13
langchain-community==0.2.6
langchain-chroma==0.1.2
langchainhub==0.1.20
chromadb==0.5.3
jq==1.7.0
streamlit
langgraph
pydantic
duckdb==1.1.1
psycopg-pool 
langgraph-checkpoint-postgres
langgraph-checkpoint-sqlite
ipykernel
notebook
duckduckgo-search
BeautifulSoup4
lxml
trafilatura==0.5.0
flask_cors
Fastapi
uvicorn
gunicorn
pydantic-settings
orjson
passlib
python-jose[cryptography]
retry
ordered-set
black